+++ src/main/java/programmers/org/voucher/repository/util/SelectBuilder.java
@@ -1,8 +1,7 @@
-package programmers.org.voucher.repository.util.sqlBuilder;
+package programmers.org.voucher.repository.util;
 
-import programmers.org.voucher.repository.util.statement.select.From;
-import programmers.org.voucher.repository.util.statement.select.Select;
-import programmers.org.voucher.repository.util.statement.Where;
+import programmers.org.voucher.repository.util.statement.From;
+import programmers.org.voucher.repository.util.statement.Select;
 
 import static programmers.org.voucher.repository.util.constant.Symbol.BLANK;
 import static programmers.org.voucher.repository.util.constant.Symbol.COMMA;
