+++ src/main/java/programmers/org/voucher/repository/util/statement/Delete.java
@@ -1,12 +1,12 @@
-package programmers.org.voucher.repository.util.statement.select;
+package programmers.org.voucher.repository.util.statement;
 
 import programmers.org.voucher.repository.util.constant.Table;
 
-public class From {
+public class Delete {
 
     private Table table;
 
-    public From(Table table) {
+    public Delete(Table table) {
         this.table = table;
     }
